Add files via upload
Исправление ошибок, добавление функционала

+++ src/main/java/com/libtask/library2/repositories/BookRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
 
     @Query(value = "SELECT * FROM book", nativeQuery = true)
     List<Book> showAllBooks();
@@ -17,6 +18,6 @@ public interface BookRepository extends JpaRepository<Book, Integer> {
     List<Book> getBooksOnBalance();
 
     @Query(value = "SELECT * FROM book WHERE user_id = :user_id", nativeQuery = true)
-    List<Book> getTakenBooksListByUserId(@Param("user_id") int userId);
+    List<Book> getTakenBooksListByUserId(@Param("user_id") Long userId);
 
 }
